+++ sympl/__init__.py
@@ -17,9 +17,10 @@ from ._core.constants import (
 from ._core.util import (
     ensure_no_shared_keys,
     get_numpy_array, jit,
-    restore_dimensions, get_numpy_arrays_with_properties,
-    restore_data_arrays_with_properties,
+    restore_dimensions,
     get_component_aliases)
+from ._core.state import (get_numpy_arrays_with_properties,
+    restore_data_arrays_with_properties)
 from ._components import (
     PlotFunctionMonitor, NetCDFMonitor, RestartMonitor,
     ConstantPrognostic, ConstantDiagnostic, RelaxationPrognostic,