+++ sympl/_core/timestepper.py
@@ -53,7 +53,7 @@ class TimeStepper(object):
     def input_properties(self):
         input_properties = combine_component_properties(
             self.prognostic_list, 'input_properties')
-        return combine_properties(input_properties, self.output_properties)
+        return combine_properties([input_properties, self.output_properties])
 
     @property
     def diagnostic_properties(self):
@@ -188,6 +188,7 @@ class TimeStepper(object):
                     'TimeStepper ({}). You must disable '
                     'tendencies_in_diagnostics for this TimeStepper.'.format(
                         tendency_name))
+            print(name, input_properties, output_properties)
             base_units = input_properties[name]['units']
             diagnostics[tendency_name] = (
                 (new_state[name].to_units(base_units) - state[name].to_units(base_units)) /